src/functions/clevertap-bq-etl/requirements.txt
# Function dependencies

google-cloud-bigquery==1.8.1
google-cloud-error-reporting


